move saveWeatherPackage to helpers
